feat: Move search field from result header box to facet column.
