+++ src/components/BiobankExplorerContainer.vue
@@ -7,7 +7,7 @@
     <div class="col-md-9">
       <div class="row">
         <div class="col-md-12">
-          <search-box></search-box>
+          <result-header></result-header>
         </div>
       </div>
 
@@ -29,7 +29,7 @@
 <script>
   import BiobankCardsContainer from './cards/BiobankCardsContainer'
   import FilterContainer from './filters/FilterContainer'
-  import SearchBox from './SearchBox'
+  import ResultHeader from './ResultHeader'
 
   import { MAP_QUERY_TO_STATE, GET_BIOBANK_IDENTIFIERS } from '../store/actions'
 
@@ -38,7 +38,7 @@
     components: {
       BiobankCardsContainer,
       FilterContainer,
-      SearchBox
+      ResultHeader
     },
     computed: {
       /* A computed 'compound' object for watching all filters */
